Move version.txt resource to core module
