+++ java/cdk/pom.xml
@@ -111,21 +111,5 @@
                 <artifactId>maven-shade-plugin</artifactId>
             </plugin>
         </plugins>
-        <resources>
-            <resource>
-                <directory>src/main/resources</directory>
-                <filtering>true</filtering>
-                <includes>
-                    <include>version.txt</include>
-                </includes>
-            </resource>
-            <resource>
-                <directory>src/main/resources</directory>
-                <filtering>false</filtering>
-                <excludes>
-                    <exclude>version.txt</exclude>
-                </excludes>
-            </resource>
-        </resources>
     </build>
 </project>
