+++ java/core/src/main/resources/version.txt
